add new AbstractView and refactoring module welcome-View
