+++ js/data/models/startGame.js
@@ -3,9 +3,7 @@ const startGame = {
   start: `Начать игру`,
   rules: {
     title: `Правила игры`,
-    description: `Правила просты&nbsp;— за&nbsp;5 минут ответить на все вопросы.`,
-    warnings: `Ошибиться можно 3 раза.`,
-    inf: `Удачи!`
+    description: `Правила просты&nbsp;— за&nbsp;5 минут ответить на все вопросы.<br>Ошибиться можно 3 раза.<br>Удачи!`,
   }
 };
